Refactoring, AB#260
