+++ src/main/java/ClientMain.java
@@ -6,7 +6,6 @@ public class ClientMain {
     public static void main(String[] args){
 
         GameEngine game = new GameEngine();
-
         game.execute();
