--- src/main/java/GameGlobalVariable.java
@@ -1,7 +1,0 @@
-package main.java;
-
-class GameGlobalVariable {
-static int currentLocationIndex = 0;
-static String currentLocation;
-
-}
